+++ src/main/java/viettel/huannt14/checklist/service/ServerInfoService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ServerInfoService {
     List<ServerInfo> getAll();
 
-    Boolean deleteById(Long serverId);
+    Boolean deleteById(Integer serverId);
 
     ServerInfo save(ServerInfo serverInfo);
