+++ src/main/java/viettel/huannt14/checklist/service_imp/ChecklistHistoryServiceImpl.java
@@ -28,7 +28,7 @@ public class ChecklistHistoryServiceImpl implements ChecklistHistoryService {
     }
 
     @Override
-    public Boolean deleteById(Long id) {
+    public Boolean deleteById(Integer id) {
         try{
             checklistHistoryRepo.deleteById(id);
             return true;
